+++ zstd.c
@@ -120,7 +120,7 @@ void zstd_module_init(PyObject* m) {
 	   We detect this mismatch here and refuse to load the module if this
 	   scenario is detected.
 	*/
-	if (ZSTD_VERSION_NUMBER != 10103 || ZSTD_versionNumber() != 10103) {
+	if (ZSTD_VERSION_NUMBER != 10104 || ZSTD_versionNumber() != 10104) {
 		PyErr_SetString(PyExc_ImportError, "zstd C API mismatch; Python bindings not compiled against expected zstd version");
 		return;
 	}