+++ tests/test_module_attributes.py
@@ -12,7 +12,7 @@ from . common import (
 @make_cffi
 class TestModuleAttributes(unittest.TestCase):
     def test_version(self):
-        self.assertEqual(zstd.ZSTD_VERSION, (1, 2, 0))
+        self.assertEqual(zstd.ZSTD_VERSION, (1, 3, 0))
 
     def test_constants(self):
         self.assertEqual(zstd.MAX_COMPRESSION_LEVEL, 22)