+++ NEWS.rst
@@ -15,6 +15,7 @@ Backwards Compatibility Notes
   been renamed to ``read_to_iter()``. ``read_from()`` is aliased to the new
   name and will be deleted in a future release.
 * Support for Python 2.6 has been removed.
+* Support for Python 3.3 has been removed.
 
 Changes
 -------