+++ NEWS.rst
@@ -38,6 +38,7 @@ Changes
   *backend* as appropriate. Behavior can be controlled via the
   ``PYTHON_ZSTANDARD_IMPORT_POLICY`` environment variable. See README for
   usage info.
+  * Vendored version of zstd upgraded to 1.3.0.
 
 0.8.1 (released 2017-04-08)
 ---------------------------